pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aigrandprix"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "opencv-python>=4.9",
    "pyyaml>=6.0",
    "pydantic>=2.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-timeout>=2.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["aigrandprix*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 120
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
